requirements.txt
flask
flask-cors
gunicorn
psycopg2-binary
python-dotenv
werkzeug
openpyxl